添加测试界面入口，完成测试版主界面
修改应用style文件（兼容新组件），添加颜色文件，导入一些包，添加了若干图标等
