+++ app/src/main/java/com/zhj/browser/ui/activity/FavActivity.kt
@@ -1,0 +1,12 @@
+package com.zhj.browser.ui.activity
+
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+
+class FavActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_fav)
+    }
+}
