+++ app/src/main/java/com/zhj/browser/ui/activity/MainActivity.kt
@@ -1,0 +1,16 @@
+package com.zhj.browser.ui.activity
+
+import android.os.Bundle
+import android.app.Activity
+import com.zhj.browser.R
+
+import kotlinx.android.synthetic.main.activity_main.*
+
+class MainActivity : Activity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+
+}
