+++ app/src/main/java/com/zhj/browser/ui/fragment/WebFragment.kt
@@ -1,0 +1,24 @@
+package com.zhj.browser.ui.fragment
+
+import android.app.Fragment
+import android.os.Bundle
+import com.tencent.smtt.sdk.WebView
+import com.tencent.smtt.sdk.WebViewClient
+import kotlinx.android.synthetic.main.fragment_web.*
+
+class WebFragment : Fragment(){
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        initUI()
+    }
+
+    private fun initUI() {
+        mWebView.webViewClient = object : WebViewClient() {
+            override fun shouldOverrideUrlLoading(view: WebView, url: String): Boolean {
+                view.loadUrl( url )
+                return true
+            }
+        }
+    }
+}
