+++ app/src/main/res/layout/activity_fav.xml
@@ -1,17 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".demo.WebViewActivity"
-    android:orientation="vertical">
+    tools:context=".ui.activity.FavActivity">
 
-    <com.tencent.smtt.sdk.WebView
-        android:id="@+id/web_view"
+    <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
+        
+    </android.support.design.widget.AppBarLayout>
 
-    </com.tencent.smtt.sdk.WebView>
-</LinearLayout>
+</android.support.design.widget.CoordinatorLayout>
